+++ src/classes/babyClasses.ts
@@ -174,6 +174,7 @@ export { LostGreyBaby } from "./babies/LostGreyBaby";
 export { LostWhiteBaby } from "./babies/LostWhiteBaby";
 export { LoveBaby } from "./babies/LoveBaby";
 export { LoveEyeBaby } from "./babies/LoveEyeBaby";
+export { LovebearBaby } from "./babies/LovebearBaby";
 export { LowfaceBaby } from "./babies/LowfaceBaby";
 export { MagBaby } from "./babies/MagBaby";
 export { MagicCatBaby } from "./babies/MagicCatBaby";